+++ src/features/basic/index.ts
@@ -34,6 +34,7 @@ import './lm-colored-buy-sell';
 import './lm-highlight-own-ads';
 import './lm-item-icons/lm-item-icons';
 import './mat-linkify-category';
+import './mtra-auto-focus-amount';
 import './nots-material-ticker';
 import './nots-notification-type-label';
 import './nots-ship-arrival-inventory';